+++ DotNet3dsToolkit/Ctr/NandNcsdHeader.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DotNet3dsToolkit
+namespace DotNet3dsToolkit.Ctr
 {
     public class NandNcsdHeader : NcsdHeader
     {
