+++ DotNet3dsToolkit/Ctr/NcsdHeader.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DotNet3dsToolkit
+namespace DotNet3dsToolkit.Ctr
 {
 
     /// <summary>
